% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprt_plot.R
\name{sprt_plot}
\alias{sprt_plot}
\title{Plot SPRT results}
\usage{
sprt_plot(res)
}
\arguments{
\item{res}{A list returned by \code{sprt()}.}
}
\value{
A ggplot object showing the SPRT path with thresholds and decision point.
}
\description{
Plot SPRT results
}
\examples{
x <- c(0,0,1,0,1,1,1,0,0,1,0,0)
res <- sprt(x, alpha = 0.05, beta = 0.1, p0 = 0.1, p1 = 0.3)
print(res)
sprt_plot(res)

x1 <- c(52, 55, 58, 63, 66, 70, 74)
result1 <- sprt(x1, alpha = 0.05, beta = 0.1, p0 = 50, p1 = 65, dist = "normal", sigma = 10)
result1
sprt_plot(result1)
}
